backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
langgraph==0.2.68
langchain-core==0.3.28
langchain-google-genai==2.0.8
langchain-nvidia-ai-endpoints==0.3.5
networkx==3.4.2
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1
mcp==1.3.0
PyMuPDF==1.25.1
aiosqlite==0.20.0
lxml==5.3.0
pytest==8.3.4
pytest-asyncio==0.25.2
